AG-10755 - convert to scale util function
