+++ packages/ag-charts-community/src/module-support.ts
@@ -53,6 +53,7 @@ export * from './chart/series/seriesProperties';
 export * from './chart/series/seriesMarker';
 export * from './chart/series/seriesTooltip';
 export * from './chart/series/seriesTypes';
+export * from './chart/series/util';
 export * from './chart/series/cartesian/abstractBarSeries';
 export * from './chart/series/cartesian/cartesianSeries';
 export * from './chart/series/cartesian/lineUtil';
